Implement futures in upload
To increase throughput, this commit implements futures, allowing
multiple files to be uploaded simultaniously.

The number of threads used and the number retry attempts allowed for
each upload can be specified when creating the Client via the
"worker_count" and "retry_count" parameters. The default for these are 4
and 3 respectively.

Modifications were also made to tests and requirements as necessary.