Implement futures in publish
To increase throughput, this commit implements futures, allowing
multiple items to be published to a table simultaniously.

The number of threads used and the number of retry attempts allowed for
each upload can be specified when creating the Client via the
"worker_count" and "retry_count" parameters. The default for these are 4
and 3 respectively.

Modifications were also made to tests and models as necessary.