Revise TableItem, Client.publish
This commit makes the TableItem class more flexible, allowing it to
represent not only an item to insert into a table but also an item
retrieved from a table.

The client's publish method was also updated to support the new
TableItem model and, as a result, recieved a more efficient means by
which it validates the the item it's publishing -- querying instead of
scanning.