Implement Client.query method
This commit implements the query method, allowing the user to query a
DynamoDB table using a TableItem as a criteria.

Included in this commit are changes to the underlying code for checking
what should be put into a table via Client.publish. The query here will
now be filtered on non-key attributes as well.

Tests were also created and updated to support these changes and the
version was increased (v2.0.0 -> v2.1.0) due to API changes.