Improve processing of dates, time in TableItems
Previously, if "date" was found in the key of a kwarg passed to
TableItem constructor, the value was forced to conform to a dateutil
parseable string. Now non-parsable values are allowed.

Additionally, support was provided for time and date-time kwargs and any
time value is converted to UTC timezone. Tests for the TableItem model
were also updated.